test(reflection): house relocated inverse extras in a trails.test sibling
Per the `*.trails.test.ts` convention, TS-only tests with no Rails analog
belong in a trails sibling rather than the convention file (where they inflate
its `extra` count — a fidelity smell). Move the 5 relocated deliberate-extra
inverse-detection tests out of inverse-associations.test.ts into a new
inverse-associations.trails.test.ts, and drop the now-unused imports from the
convention file. The scope markers use `unknown` rather than `any` so the new
file need not join the no-explicit-any test-exclude allowlist.

inverse_associations_test.rb extra returns to its baseline (2, not 7); matched
unchanged at 92. reflection_test.rb still: extra down by 8, matched 63.
